refactor: 디렉토리 구조 리팩토링
개발 환경에서 include, src 두개 디렉토리만 사용하도록 디렉토리 구조 변경
미구현 헤더파일 내용 삭제
변경 내용 CMakeLists.txt반영, 빌드 테스트
