+++ server/include/NetworkManager.h
