+++ server/include/ServerTypes.h
