+++ server/src/GameRoom.cpp
@@ -1,5 +1,5 @@
-﻿#include "manager/GameRoom.h"
-#include "core/Session.h"
+﻿#include "GameRoom.h"
+#include "Session.h"
 #include <spdlog/spdlog.h>
 #include <algorithm>
 #include <sstream>
