+++ server/src/session.cpp
@@ -1,5 +1,5 @@
-﻿#include "core/Session.h"
-#include "handler/MessageHandler.h"
+﻿#include "Session.h"
+#include "MessageHandler.h"
 #include <openssl/rand.h>
 #include <chrono>
 #include <iomanip>
